+++ packages/maskbook/src/social-network-adaptor/twitter.com/injection/EnhancedProfileTab.tsx
@@ -1,0 +1,31 @@
+import { MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
+import { EnhancedProfileaPage, EnhancedProfileTab } from '../../../components/InjectedComponents/EnhancedProfile'
+import { createReactRootShadowed, startWatch } from '../../../utils'
+import {
+    searchProfileEmptySelector,
+    searchProfileTabListLastChildSelector,
+    searchProfileTabPageSelector,
+} from '../utils/selector'
+
+function injectEnhancedProfileTab(signal: AbortSignal) {
+    const watcher = new MutationObserverWatcher(searchProfileTabListLastChildSelector())
+    startWatch(watcher, signal)
+    createReactRootShadowed(watcher.firstDOMProxy.afterShadow, { signal }).render(<EnhancedProfileTab />)
+}
+
+function injectEnhancedProfilePageForEmptyState(signal: AbortSignal) {
+    const watcher = new MutationObserverWatcher(searchProfileEmptySelector())
+    startWatch(watcher, signal)
+    createReactRootShadowed(watcher.firstDOMProxy.afterShadow, { signal }).render(<EnhancedProfileaPage />)
+}
+
+function injectEnhancedProfilePageState(signal: AbortSignal) {
+    const watcher = new MutationObserverWatcher(searchProfileTabPageSelector())
+    startWatch(watcher, signal)
+    createReactRootShadowed(watcher.firstDOMProxy.afterShadow, { signal }).render(<EnhancedProfileaPage />)
+}
+export function injectEnhancedProfile(signal: AbortSignal) {
+    injectEnhancedProfileTab(signal)
+    injectEnhancedProfilePageForEmptyState(signal)
+    injectEnhancedProfilePageState(signal)
+}
